// public/angular/app/interfaces/IComparendo.ts
export interface IComparendo {
    comparendo: number;
    deteccion: number;
    numero: string;
    infractor: number;
    infr_tipo_doc: number;
    infr_tipo_doc_desc: string;
    infr_numero_doc: string;
    infr_nombres: string;
    infr_apellidos: string;
    infr_nombres_apellidos: string;
    divipo_infractor: number;
    dinfr_cod_departamento: number;
    dinfr_departamento: string;
    dinfr_cod_municipio: number;
    dinfr_municipio: string;
    direccion_infractor: string;
    telefono_infractor: string;
    email_infractor: string;
    edad_infractor: string;
    lcond_numero: string;
    lcond_categoria: number;
    lcond_categoria_desc: string;
    lcond_expedicion: string;
    lcond_vencimiento: string;
    lcond_organismo: string;
    agente: number;
    agt_placa: string;
    agt_entidad: number;
    agt_entidad_desc: string;
    agt_tipo_doc: number;
    agt_tipo_doc_desc: string;
    agt_numero_doc: string;
    agt_nombres: string;
    agt_apellidos: string;
    agt_nombres_apellidos: string;
    agt_tiene_firma: number;
    agt_firma: string;
    infraccion: number;
    infr_codigo: string;
    infr_nombre_corto: string;
    infr_valor: number;
    fecha_deteccion: string;
    fecha_imposicion: string;
    divipo: number;
    div_cod_departamento: number;
    div_departamento: string;
    div_cod_municipio: number;
    div_municipio: string;
    direccion: string;
    longitud: number;
    latitud: number;
    placa_vehiculo: string;
    clase_vehiculo: number;
    clase_vehiculo_desc: string;
    servicio_vehiculo: number;
    servicio_vehiculo_desc: string;
    organismo_vehiculo: number;
    organismo_vehiculo_desc: string;
    licencia_vehiculo: string;
    propietario_vehiculo: number;
    prop_tipo_doc: number;
    prop_tipo_doc_desc: string;
    prop_numero_doc: string;
    prop_nombres: string;
    prop_apellidos: string;
    prop_nombres_apellidos: string;
    polca: number;
    estado: number;
    estado_desc: string;
    etapa_proceso: number;
    etapa_proceso_desc: string;
    cartera: number;
    inmovilizado: number;
    observaciones: string;
    nit_empresa_tte: string;
    nombre_empresa: string;
    tarjeta_operacion: string;
    modalidad: number;
    modalidadd_desc: string;
    radio_accion: number;
    radio_accion_desc: string;
    tipo_pasajero: number;
    tipo_pasajero_desc: string;
    impreso: number;
}