// public/angular/app/interfaces/IPersona.ts
export interface IPersona {
    persona: number;
    tipo_doc: number;
    tipo_doc_desc: string;
    numero_doc: string;
    fecha_exped_doc: string;
    divipo_doc: number;
    cod_departamento_doc: number;
    cod_municipio_doc: number;
    nombres: string;
    apellidos: string;
    nombres_apellidos: string;
    email: string;
    genero: number;
    genero_desc: string;
    grupo_sanguineo: number;
    grupo_sanguineo_desc: string;
    numero_celular: number;
    fecha_registro: string;
    usuario_registra: number;
    usuario_registra_desc: string;
}